// projects/ui-builder-component-lib/src/lib/enums/form-field-type.ts
export enum FormFieldType {
  date = 'date',
  text = 'text',
  textarea = 'textarea',
  typeahead = 'typeahead',
  multiselect = 'multiselect',
  select = 'select',
  checkbox = 'checkbox',
  toggle = 'toggle',
  radio = 'radio',
  textlist = 'textlist',
  number = 'number',
  message = 'message',
  time = 'time',
  multifield = 'multifield',
  action = 'action',
  currency = 'currency',
  banner = 'banner',
  percentage = 'percentage',
  multiselectdropdown = 'multiselectdropdown',
  hyperlink = 'hyperlink',
  injectable = 'injectable',
  datetime = 'datetime',
  rating = 'rating',
  incrementer = 'incrementer',
  range = 'range',
  editor = 'editor',
}